Turned the unittest tests into pytest tests. Localy changed Pycharm default test runner to Pycharm to make things work. All tests currently passing
